Divide ActiveRecord::ConnectionAdapters::OracleEnhancedColumnDumper and ActiveRecord::ConnectionAdapters::OracleEnhancedSchemaDumper
To prepare merge rails4 branch into master,
these modules should be stored in separe file because
ActiveRecord::ConnectionAdapters::OracleEnhancedColumnDumper just supports
Rails 4 and does not Rails 3.2 and earlier versions.
